+++ hotel_management_odoo/models/__init__.py
@@ -29,10 +29,10 @@ from . import fleet_vehicle_model
 from . import food_booking_line
 from . import hotel_amenity
 from . import hotel_floor
-from . import hotel_room
 from . import hotel_service
 from . import maintenance_request
 from . import maintenance_team
 from . import room_booking
 from . import room_booking_line
 from . import service_booking_line
+from . import product_template