+++ hotel_management_odoo/models/cleaning_request.py
@@ -19,7 +19,7 @@
 #    If not, see <http://www.gnu.org/licenses/>.
 #
 #############################################################################
-from odoo import api, fields, models
+from odoo import api, fields, models, _
 from odoo.exceptions import ValidationError
 
 
@@ -46,7 +46,7 @@ class CleaningRequest(models.Model):
                                      required=True, tracking=True,
                                      string="Cleaning Type",
                                      help="Choose what is to be cleaned")
-    room_id = fields.Many2one('hotel.room', string="Room",
+    room_id = fields.Many2one('product.template', string="Room",
                               help="Choose the room")
     hotel = fields.Char(string="Hotel", help="Cleaning request space in hotel")
     vehicle_id = fields.Many2one('fleet.vehicle.model',
@@ -97,6 +97,13 @@ class CleaningRequest(models.Model):
         """Button action for updating the state to assign"""
         if not self.assigned_id:
             raise ValidationError(_('Please select an Assigned To person before assigning.'))
+        if self.cleaning_type == 'room' and not self.room_id:
+            raise ValidationError(_('Please choose a Room.'))
+        if self.cleaning_type == 'hotel' and not self.hotel:
+            raise ValidationError(_('Please enter the Hotel.'))
+        if self.cleaning_type == 'vehicle' and not self.vehicle_id:
+            raise ValidationError(_('Please choose a Vehicle.'))
+
         self.update({'state': 'assign'})
 
     def action_start_cleaning(self):
@@ -123,14 +130,20 @@ class CleaningRequest(models.Model):
 
     def action_maintain_request(self):
         """Button action for creating the maintenance request"""
-        self.env['maintenance.request'].sudo().create({
-            'name': 'Maintenance Request - %s' % self.name,
+        vals = {
+            'name': 'Maintenance Request - %s' % self.sequence,
             'date': fields.Date.today(),
             'state': 'draft',
             'type': self.cleaning_type,
-            'vehicle_maintenance_id': self.vehicle_id.id,
             'is_hotel': True,
-        })
+        }
+        if self.cleaning_type == 'room':
+            vals['room_maintenance_ids'] = [(6, 0, self.room_id.ids)]
+        elif self.cleaning_type == 'vehicle':
+            vals['vehicle_maintenance_id'] = self.vehicle_id.id
+        elif self.cleaning_type == 'hotel':
+            vals['hotel_maintenance'] = self.hotel
+        self.env['maintenance.request'].sudo().create(vals)
         return {
             'type': 'ir.actions.client',
             'tag': 'display_notification',