+++ hotel_management_odoo/models/maintenance_request.py
@@ -30,7 +30,7 @@ class MaintenanceRequest(models.Model):
     _rec_name = 'sequence'
     _description = "Maintenance Request"
 
-    is_hotel = fields.Boolean(string="Is Hotel Maintenance", default=False)
+    is_hotel = fields.Boolean(string="Is Hotel Maintenance", default=False,help='is hotel')
 
     sequence = fields.Char(readonly=True, string="Sequence", copy=False,
                            default='New', help='Sequence number for'
@@ -70,7 +70,7 @@ class MaintenanceRequest(models.Model):
                                        ('cleaning', 'Cleaning')], string="Type",
                             help="The type for which the request is creating",
                             tracking=True)
-    room_maintenance_ids = fields.Many2many('hotel.room',
+    room_maintenance_ids = fields.Many2many('product.template',
                                             string="Room Maintenance",
                                             help="Choose Room Maintenance")
     hotel_maintenance = fields.Char(string='Hotel Maintenance',
@@ -89,8 +89,9 @@ class MaintenanceRequest(models.Model):
     support_reason = fields.Char(string='Support',
                                  help="Reason for adding Support")
     remarks = fields.Char(string='Remarks', help="Add Remarks")
-    domain_partner_ids = fields.Many2many('res.partner',
+    domain_partner_ids = fields.Many2many('res.users',
                                           string="Partner",
+                                          compute='_compute_domain_partner_ids',
                                           help="For filtering Users")
 
     @api.model_create_multi
@@ -100,17 +101,28 @@ class MaintenanceRequest(models.Model):
             if vals.get('sequence', 'New') == 'New':
                 vals['sequence'] = self.env['ir.sequence'].next_by_code(
                     'maintenance.request')
+            if not vals.get('name'):
+                vals['name'] = vals['sequence']
         return super().create(vals_list)
 
-    @api.onchange('team_id')
-    def _onchange_team_id(self):
+    @api.depends('team_id')
+    def _compute_domain_partner_ids(self):
         """Function for filtering the maintenance team user"""
-        self.update({
-            'domain_partner_ids': self.team_id.member_ids.ids
-        })
+        for record in self:
+            record.domain_partner_ids = record.team_id.member_ids
 
     def action_assign_team(self):
         """Button action for changing the state to team_leader_approve"""
+        if self.type == 'room' and not self.room_maintenance_ids:
+            raise ValidationError(("Please choose the Room"))
+        if self.type == 'vehicle' and not self.vehicle_maintenance_id:
+            raise ValidationError(("Please choose the Vehicle"))
+        if self.type == 'hotel' and not self.hotel_maintenance:
+            raise ValidationError(
+                ("Please enter the Hotel Maintenance details"))
+        if self.type == 'cleaning' and not self.cleaning_maintenance:
+            raise ValidationError(
+                ("Please enter the Cleaning Maintenance details"))
         if self.team_id:
             self.state = 'team_leader_approve'
         else:
@@ -141,7 +153,7 @@ class MaintenanceRequest(models.Model):
         if self.remarks:
             self.state = 'verify'
         else:
-            raise ValidationError(_('Please Add remark'))
+            raise ValidationError('Please Add remark')
 
     def action_assign_support(self):
         """Button action for changing the state to ongoing"""