+++ hotel_management_odoo/doc/RELEASE_NOTES.md
@@ -10,3 +10,13 @@
 #### Version 19.0.1.0.1
 #### UPDT
 - Fixed menu action issue during module installation.
+
+#### 06.07.2026
+#### Version 19.0.1.0.2
+#### UPDT
+- Fixed menu validation,maintenence request,cleaning request issues .
+
+#### 08.07.2026
+#### Version 19.0.1.0.3
+#### UPDT
+- Fixed the room workflow by changing the room model to use product.template and updated the Sales Report by adding the state field.